*** empty log message ***

+++ src/contrib/euler/ang.h
@@ -1,3 +1,3 @@
 
       COMMON /ANG/ ALPHA,BETA,PHI
-      double precision ALPHA,BETA,PHI
+      Double ALPHA,BETA,PHI
