+++ src/contrib/euler/avctrs.h
@@ -1,18 +1,18 @@
 c
 c Parallel mesh metrics.
-      double precision aix(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
-      double precision ajx(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
-      double precision akx(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
-      double precision aiy(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
-      double precision ajy(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
-      double precision aky(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
-      double precision aiz(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
-      double precision ajz(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
-      double precision akz(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double aix(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double ajx(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double akx(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double aiy(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double ajy(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double aky(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double aiz(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double ajz(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
+      Double akz(gxsf1:gxef01,gysf1:gyef01,gzsf1:gzef01)
 
 c Uniprocessor version:
 c      COMMON /AVCTRS/ AIX(NI,NJ,NK),AJX(NI,NJ,NK),AKX(NI,NJ,NK)
 c      COMMON /AVCTRS/ AIY(NI,NJ,NK),AJY(NI,NJ,NK),AKY(NI,NJ,NK)
 c      COMMON /AVCTRS/ AIZ(NI,NJ,NK),AJZ(NI,NJ,NK),AKZ(NI,NJ,NK)
-c      double precision  AIX,AJX,AKX,AIY,AJY,AKY,AIZ,AJZ,AKZ
+c      Double  AIX,AJX,AKX,AIY,AJY,AKY,AIZ,AJZ,AKZ
 c
