+++ src/contrib/euler/sigma.h
@@ -1,9 +1,9 @@
 c
-      double precision SP(5,ni),SM(5,ni)
-      double precision SP1(5,ni),SP2(5,ni)
-      double precision SM1(5,ni),SM2(5,ni)
+      Double SP(5,ni),SM(5,ni)
+      Double SP1(5,ni),SP2(5,ni)
+      Double SM1(5,ni),SM2(5,ni)
 
 c      COMMON /SIGMA/ SP(5,NI),SM(5,NI)
 c      COMMON /SIGMA/ SP1(5,NI),SP2(5,NI),SM1(5,NI),SM2(5,NI)
-c      double precision sp,sm,sp1,sp2,sm1,sm2
+c      Double sp,sm,sp1,sp2,sm1,sm2
